+++ packages/twenty-front/src/modules/object-record/record-field-list/components/RecordFieldList.tsx
@@ -1,0 +1,250 @@
+import { ActivityTargetsInlineCell } from '@/activities/inline-cell/components/ActivityTargetsInlineCell';
+import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
+import { type CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
+import { formatFieldMetadataItemAsColumnDefinition } from '@/object-metadata/utils/formatFieldMetadataItemAsColumnDefinition';
+import { useObjectPermissions } from '@/object-record/hooks/useObjectPermissions';
+import { RecordFieldListCellEditModePortal } from '@/object-record/record-field-list/anchored-portal/components/RecordFieldListCellEditModePortal';
+import { RecordFieldListCellHoveredPortal } from '@/object-record/record-field-list/anchored-portal/components/RecordFieldListCellHoveredPortal';
+import { useFieldListFieldMetadataItems } from '@/object-record/record-field-list/hooks/useFieldListFieldMetadataItems';
+import { RecordDetailDuplicatesSection } from '@/object-record/record-field-list/record-detail-section/duplicate/components/RecordDetailDuplicatesSection';
+import { RecordDetailRelationSection } from '@/object-record/record-field-list/record-detail-section/relation/components/RecordDetailRelationSection';
+import { RecordFieldListComponentInstanceContext } from '@/object-record/record-field-list/states/contexts/RecordFieldListComponentInstanceContext';
+import { recordFieldListHoverPositionComponentState } from '@/object-record/record-field-list/states/recordFieldListHoverPositionComponentState';
+import { FieldContext } from '@/object-record/record-field/ui/contexts/FieldContext';
+import { useIsRecordReadOnly } from '@/object-record/record-field/ui/hooks/read-only/useIsRecordReadOnly';
+import { isRecordFieldReadOnly } from '@/object-record/record-field/ui/hooks/read-only/utils/isRecordFieldReadOnly';
+import { RecordFieldComponentInstanceContext } from '@/object-record/record-field/ui/states/contexts/RecordFieldComponentInstanceContext';
+import { RecordInlineCell } from '@/object-record/record-inline-cell/components/RecordInlineCell';
+import { PropertyBox } from '@/object-record/record-inline-cell/property-box/components/PropertyBox';
+import { PropertyBoxSkeletonLoader } from '@/object-record/record-inline-cell/property-box/components/PropertyBoxSkeletonLoader';
+import { useRecordShowContainerActions } from '@/object-record/record-show/hooks/useRecordShowContainerActions';
+import { useRecordShowContainerData } from '@/object-record/record-show/hooks/useRecordShowContainerData';
+import { getRecordFieldInputInstanceId } from '@/object-record/utils/getRecordFieldInputId';
+import { getObjectPermissionsFromMapByObjectMetadataId } from '@/settings/roles/role-permissions/objects-permissions/utils/getObjectPermissionsFromMapByObjectMetadataId';
+import { useIsInRightDrawerOrThrow } from '@/ui/layout/right-drawer/contexts/RightDrawerContext';
+import { useSetRecoilComponentState } from '@/ui/utilities/state/component-state/hooks/useSetRecoilComponentState';
+
+type RecordFieldListProps = {
+  objectNameSingular: string;
+  objectRecordId: string;
+  showDuplicatesSection?: boolean;
+};
+
+const INPUT_ID_PREFIX = 'fields-card';
+
+export const RecordFieldList = ({
+  objectNameSingular,
+  objectRecordId,
+  showDuplicatesSection = true,
+}: RecordFieldListProps) => {
+  const { recordLoading, isPrefetchLoading } = useRecordShowContainerData({
+    objectRecordId,
+  });
+
+  const { objectMetadataItem } = useObjectMetadataItem({
+    objectNameSingular,
+  });
+
+  const { objectPermissionsByObjectMetadataId } = useObjectPermissions();
+
+  const { useUpdateOneObjectRecordMutation } = useRecordShowContainerActions({
+    objectNameSingular,
+    objectRecordId,
+  });
+
+  const { isInRightDrawer } = useIsInRightDrawerOrThrow();
+
+  const isRecordReadOnly = useIsRecordReadOnly({
+    recordId: objectRecordId,
+    objectMetadataId: objectMetadataItem.id,
+  });
+
+  const setRecordFieldListHoverPosition = useSetRecoilComponentState(
+    recordFieldListHoverPositionComponentState,
+    `fields-list-${objectRecordId}`,
+  );
+
+  const handleMouseEnter = (index: number) => {
+    setRecordFieldListHoverPosition(index);
+  };
+
+  const {
+    inlineFieldMetadataItems,
+    inlineRelationFieldMetadataItems,
+    boxedRelationFieldMetadataItems,
+  } = useFieldListFieldMetadataItems({
+    objectNameSingular,
+  });
+
+  return (
+    <RecordFieldListComponentInstanceContext.Provider
+      value={{
+        instanceId: `fields-list-${objectRecordId}`,
+      }}
+    >
+      <PropertyBox>
+        {isPrefetchLoading ? (
+          <PropertyBoxSkeletonLoader />
+        ) : (
+          <>
+            {inlineRelationFieldMetadataItems?.map(
+              (fieldMetadataItem, index) => (
+                <FieldContext.Provider
+                  key={objectRecordId + fieldMetadataItem.id}
+                  value={{
+                    recordId: objectRecordId,
+                    maxWidth: 200,
+                    isLabelIdentifier: false,
+                    fieldDefinition: formatFieldMetadataItemAsColumnDefinition({
+                      field: fieldMetadataItem,
+                      position: index,
+                      objectMetadataItem,
+                      showLabel: true,
+                      labelWidth: 90,
+                    }),
+                    useUpdateRecord: useUpdateOneObjectRecordMutation,
+                    isDisplayModeFixHeight: true,
+                    onMouseEnter: () => handleMouseEnter(index),
+                    anchorId: `record-field-list-inline-cell-${index}`,
+                    isRecordFieldReadOnly: isRecordFieldReadOnly({
+                      isRecordReadOnly,
+                      objectPermissions:
+                        getObjectPermissionsFromMapByObjectMetadataId({
+                          objectPermissionsByObjectMetadataId,
+                          objectMetadataId: objectMetadataItem.id,
+                        }),
+                      fieldMetadataId: fieldMetadataItem.id,
+                      objectNameSingular,
+                      fieldName: fieldMetadataItem.name,
+                      fieldType: fieldMetadataItem.type,
+                      isCustom: fieldMetadataItem.isCustom ?? false,
+                    }),
+                  }}
+                >
+                  <ActivityTargetsInlineCell
+                    componentInstanceId={getRecordFieldInputInstanceId({
+                      recordId: objectRecordId,
+                      fieldName: fieldMetadataItem.name,
+                      prefix: isInRightDrawer
+                        ? 'right-drawer-fields-card'
+                        : 'fields-card',
+                    })}
+                    activityObjectNameSingular={
+                      objectNameSingular as
+                        | CoreObjectNameSingular.Note
+                        | CoreObjectNameSingular.Task
+                    }
+                    activityRecordId={objectRecordId}
+                    showLabel={true}
+                    maxWidth={200}
+                  />
+                </FieldContext.Provider>
+              ),
+            )}
+            {inlineFieldMetadataItems?.map((fieldMetadataItem, index) => (
+              <FieldContext.Provider
+                key={objectRecordId + fieldMetadataItem.id}
+                value={{
+                  recordId: objectRecordId,
+                  maxWidth: 200,
+                  isLabelIdentifier: false,
+                  fieldDefinition: formatFieldMetadataItemAsColumnDefinition({
+                    field: fieldMetadataItem,
+                    position: index,
+                    objectMetadataItem,
+                    showLabel: true,
+                    labelWidth: 90,
+                  }),
+                  useUpdateRecord: useUpdateOneObjectRecordMutation,
+                  isDisplayModeFixHeight: true,
+                  isRecordFieldReadOnly: isRecordFieldReadOnly({
+                    isRecordReadOnly,
+                    objectPermissions:
+                      getObjectPermissionsFromMapByObjectMetadataId({
+                        objectPermissionsByObjectMetadataId,
+                        objectMetadataId: objectMetadataItem.id,
+                      }),
+                    fieldMetadataId: fieldMetadataItem.id,
+                    objectNameSingular,
+                    fieldName: fieldMetadataItem.name,
+                    fieldType: fieldMetadataItem.type,
+                    isCustom: fieldMetadataItem.isCustom ?? false,
+                  }),
+                  onMouseEnter: () =>
+                    handleMouseEnter(
+                      index + (inlineRelationFieldMetadataItems?.length ?? 0),
+                    ),
+                  anchorId: `record-field-list-inline-cell-${
+                    index + (inlineRelationFieldMetadataItems?.length ?? 0)
+                  }`,
+                }}
+              >
+                <RecordFieldComponentInstanceContext.Provider
+                  value={{
+                    instanceId: getRecordFieldInputInstanceId({
+                      recordId: objectRecordId,
+                      fieldName: fieldMetadataItem.name,
+                      prefix: INPUT_ID_PREFIX,
+                    }),
+                  }}
+                >
+                  <RecordInlineCell
+                    loading={recordLoading}
+                    instanceIdPrefix={INPUT_ID_PREFIX}
+                  />
+                </RecordFieldComponentInstanceContext.Provider>
+              </FieldContext.Provider>
+            ))}
+          </>
+        )}
+      </PropertyBox>
+      {showDuplicatesSection && (
+        <RecordDetailDuplicatesSection
+          objectRecordId={objectRecordId}
+          objectNameSingular={objectNameSingular}
+        />
+      )}
+      {boxedRelationFieldMetadataItems?.map((fieldMetadataItem, index) => (
+        <FieldContext.Provider
+          key={objectRecordId + fieldMetadataItem.id}
+          value={{
+            recordId: objectRecordId,
+            isLabelIdentifier: false,
+            fieldDefinition: formatFieldMetadataItemAsColumnDefinition({
+              field: fieldMetadataItem,
+              position: index,
+              objectMetadataItem,
+            }),
+            useUpdateRecord: useUpdateOneObjectRecordMutation,
+            isDisplayModeFixHeight: true,
+            isRecordFieldReadOnly: isRecordFieldReadOnly({
+              isRecordReadOnly,
+              objectPermissions: getObjectPermissionsFromMapByObjectMetadataId({
+                objectPermissionsByObjectMetadataId,
+                objectMetadataId: objectMetadataItem.id,
+              }),
+              fieldMetadataId: fieldMetadataItem.id,
+              objectNameSingular,
+              fieldName: fieldMetadataItem.name,
+              fieldType: fieldMetadataItem.type,
+              isCustom: fieldMetadataItem.isCustom ?? false,
+            }),
+          }}
+        >
+          <RecordDetailRelationSection
+            loading={isPrefetchLoading || recordLoading}
+          />
+        </FieldContext.Provider>
+      ))}
+
+      <RecordFieldListCellHoveredPortal
+        objectMetadataItem={objectMetadataItem}
+        recordId={objectRecordId}
+      />
+      <RecordFieldListCellEditModePortal
+        objectMetadataItem={objectMetadataItem}
+        recordId={objectRecordId}
+      />
+    </RecordFieldListComponentInstanceContext.Provider>
+  );
+};
