+++ packages/twenty-front/src/modules/object-record/record-field-list/hooks/useRecordDetailSectionFieldMetadataItems.ts
