+++ packages/twenty-front/src/modules/object-record/record-field-list/record-detail-section/components/RecordDetailRecordsListContainer.tsx
@@ -4,4 +4,4 @@ const StyledRecordsList = styled.div`
   color: ${({ theme }) => theme.font.color.secondary};
 `;
 
-export { StyledRecordsList as RecordDetailRecordsList };
+export { StyledRecordsList as RecordDetailRecordsListContainer };
