+++ packages/twenty-front/src/modules/object-record/record-field-list/record-detail-section/components/RecordDetailSectionContainer.tsx
@@ -1,6 +1,14 @@
 import styled from '@emotion/styled';
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
+import { Section } from 'twenty-ui/layout';
+
+const StyledRecordDetailSectionContainer = styled(Section)`
+  border-top: 1px solid ${({ theme }) => theme.border.color.light};
+  padding: ${({ theme }) => theme.spacing(3)};
+  padding-right: ${({ theme }) => theme.spacing(2)};
+  width: auto;
+`;
 
 const StyledHeader = styled.header<{
   isDropdownOpen?: boolean;
@@ -34,7 +42,8 @@ const StyledLink = styled(Link)`
   }
 `;
 
-type RecordDetailSectionHeaderProps = {
+type RecordDetailSectionContainerProps = {
+  children: React.ReactNode;
   title: string;
   link?: { to: string; label: string };
   rightAdornment?: React.ReactNode;
@@ -42,28 +51,31 @@ type RecordDetailSectionHeaderProps = {
   areRecordsAvailable?: boolean;
 };
 
-export const RecordDetailSectionHeader = ({
+export const RecordDetailSectionContainer = ({
+  children,
   title,
   link,
   rightAdornment,
   hideRightAdornmentOnMouseLeave = true,
   areRecordsAvailable = false,
-}: RecordDetailSectionHeaderProps) => {
+}: RecordDetailSectionContainerProps) => {
   const [isHovered, setIsHovered] = useState(false);
-
   return (
-    <StyledHeader
-      areRecordsAvailable={areRecordsAvailable}
-      onMouseEnter={() => setIsHovered(true)}
-      onMouseLeave={() => setIsHovered(false)}
-    >
-      <StyledTitle>
-        <StyledTitleLabel>{title}</StyledTitleLabel>
-        {link && <StyledLink to={link.to}>{link.label}</StyledLink>}
-      </StyledTitle>
-      {hideRightAdornmentOnMouseLeave && !isHovered && areRecordsAvailable
-        ? null
-        : rightAdornment}
-    </StyledHeader>
+    <StyledRecordDetailSectionContainer>
+      <StyledHeader
+        areRecordsAvailable={areRecordsAvailable}
+        onMouseEnter={() => setIsHovered(true)}
+        onMouseLeave={() => setIsHovered(false)}
+      >
+        <StyledTitle>
+          <StyledTitleLabel>{title}</StyledTitleLabel>
+          {link && <StyledLink to={link.to}>{link.label}</StyledLink>}
+        </StyledTitle>
+        {hideRightAdornmentOnMouseLeave && !isHovered && areRecordsAvailable
+          ? null
+          : rightAdornment}
+      </StyledHeader>
+      {children}
+    </StyledRecordDetailSectionContainer>
   );
 };
