+++ packages/twenty-front/src/modules/object-record/record-field-list/record-detail-section/duplicate/components/RecordDetailDuplicatesSection.tsx
@@ -2,10 +2,9 @@ import { useOpenMergeRecordsPageInCommandMenu } from '@/command-menu/hooks/useOp
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
 import { RecordChip } from '@/object-record/components/RecordChip';
 import { useFindDuplicateRecords } from '@/object-record/hooks/useFindDuplicateRecords';
-import { RecordDetailRecordsList } from '@/object-record/record-show/record-detail-section/components/RecordDetailRecordsList';
-import { RecordDetailRecordsListItem } from '@/object-record/record-show/record-detail-section/components/RecordDetailRecordsListItem';
-import { RecordDetailSection } from '@/object-record/record-show/record-detail-section/components/RecordDetailSection';
-import { RecordDetailSectionHeader } from '@/object-record/record-show/record-detail-section/components/RecordDetailSectionHeader';
+import { RecordDetailRecordsListContainer } from '@/object-record/record-field-list/record-detail-section/components/RecordDetailRecordsListContainer';
+import { RecordDetailRecordsListItemContainer } from '@/object-record/record-field-list/record-detail-section/components/RecordDetailRecordsListItemContainer';
+import { RecordDetailSectionContainer } from '@/object-record/record-field-list/record-detail-section/components/RecordDetailSectionContainer';
 
 import { isDefined } from 'twenty-shared/utils';
 import { IconArrowMerge } from 'twenty-ui/display';
@@ -44,28 +43,27 @@ export const RecordDetailDuplicatesSection = ({
     return null;
 
   return (
-    <RecordDetailSection>
-      <RecordDetailSectionHeader
-        title="Duplicates"
-        rightAdornment={
-          <LightIconButton
-            className="displayOnHover"
-            Icon={IconArrowMerge}
-            accent="tertiary"
-            onClick={openMergeRecordsPageInCommandMenu}
-          />
-        }
-      />
-      <RecordDetailRecordsList>
+    <RecordDetailSectionContainer
+      title="Duplicates"
+      rightAdornment={
+        <LightIconButton
+          className="displayOnHover"
+          Icon={IconArrowMerge}
+          accent="tertiary"
+          onClick={openMergeRecordsPageInCommandMenu}
+        />
+      }
+    >
+      <RecordDetailRecordsListContainer>
         {queryResults[0].slice(0, 5).map((duplicateRecord) => (
-          <RecordDetailRecordsListItem key={duplicateRecord.id}>
+          <RecordDetailRecordsListItemContainer key={duplicateRecord.id}>
             <RecordChip
               record={duplicateRecord}
               objectNameSingular={objectNameSingular}
             />
-          </RecordDetailRecordsListItem>
+          </RecordDetailRecordsListItemContainer>
         ))}
-      </RecordDetailRecordsList>
-    </RecordDetailSection>
+      </RecordDetailRecordsListContainer>
+    </RecordDetailSectionContainer>
   );
 };
