--- packages/twenty-front/src/modules/object-record/record-show/record-detail-section/anchored-portal/components/RecordDetailSectionCellEditModePortalContent.tsx
@@ -1,10 +1,0 @@
-import { FieldInput } from '@/object-record/record-field/ui/components/FieldInput';
-import { RecordInlineCellEditMode } from '@/object-record/record-inline-cell/components/RecordInlineCellEditMode';
-
-export const RecordDetailSectionCellEditModePortalContent = () => {
-  return (
-    <RecordInlineCellEditMode>
-      <FieldInput />
-    </RecordInlineCellEditMode>
-  );
-};
