--- packages/twenty-front/src/modules/object-record/record-show/record-detail-section/states/contexts/RecordDetailSectionComponentInstanceContext.ts
@@ -1,4 +1,0 @@
-import { createComponentInstanceContext } from '@/ui/utilities/state/component-state/utils/createComponentInstanceContext';
-
-export const RecordDetailSectionComponentInstanceContext =
-  createComponentInstanceContext();
