--- packages/twenty-front/src/modules/object-record/record-show/record-detail-section/states/recordDetailSectionHoverPositionComponentState.ts
@@ -1,9 +1,0 @@
-import { RecordDetailSectionComponentInstanceContext } from '@/object-record/record-show/record-detail-section/states/contexts/RecordDetailSectionComponentInstanceContext';
-import { createComponentState } from '@/ui/utilities/state/component-state/utils/createComponentState';
-
-export const recordDetailSectionHoverPositionComponentState =
-  createComponentState<number | null>({
-    key: 'recordDetailSectionHoverPositionComponentState',
-    defaultValue: null,
-    componentInstanceContext: RecordDetailSectionComponentInstanceContext,
-  });
